Немного изменил структуру проекта
