+++ .gitignore
@@ -1,1 +1,5 @@
 .env
+__pycache__
+.idea
+.vscode
+.venv
