+++ alembic/versions/f5819c1880eb_поменял_название_таблицы.py
@@ -1,0 +1,43 @@
+"""поменял название таблицы
+
+Revision ID: f5819c1880eb
+Revises: 3a63e3dbdf02
+Create Date: 2025-08-19 14:27:37.418038
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'f5819c1880eb'
+down_revision: Union[str, Sequence[str], None] = '3a63e3dbdf02'
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('clients',
+    sa.Column('telegram_id', sa.BigInteger(), nullable=False),
+    sa.Column('username', sa.String(), nullable=False),
+    sa.Column('language', sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint('telegram_id')
+    )
+    op.drop_table('users')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('users',
+    sa.Column('telegram_id', sa.BIGINT(), autoincrement=False, nullable=False),
+    sa.Column('username', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('language', sa.VARCHAR(), autoincrement=False, nullable=True)
+    )
+    op.drop_table('clients')
+    # ### end Alembic commands ###
