--- crud/engine.py
@@ -1,7 +1,0 @@
-
-from sqlalchemy.ext.asyncio import create_async_engine, async_sessionmaker, AsyncSession
-
-from config import settings
-
-engine = create_async_engine(settings.DB_URL, echo=True)
-session_maker = async_sessionmaker(bind=engine, class_=AsyncSession, expire_on_commit=False)
