+++ src/core/settings.py
@@ -1,0 +1,11 @@
+from pydantic_settings import BaseSettings, SettingsConfigDict
+
+
+class Settings(BaseSettings):
+    BOT_TOKEN: str
+    # ADMIN_ID: int
+    DB_URL: str
+    model_config = SettingsConfigDict(env_file=".env")
+
+
+settings = Settings()
