+++ src/database/models/models.py
@@ -1,0 +1,14 @@
+from sqlalchemy.orm import Mapped, mapped_column, DeclarativeBase
+from sqlalchemy import BigInteger
+
+class Base(DeclarativeBase):
+    pass
+
+class User(Base):
+    __tablename__ = "clients"
+
+    telegram_id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
+    username: Mapped[str] = mapped_column(nullable=False)
+    language: Mapped[str] = mapped_column(nullable=True)
+
+#TODO: создать услугу или товар
