+++ src/middlewares/db_session.py
@@ -1,0 +1,21 @@
+from typing import Awaitable, Callable, Any
+
+from aiogram import BaseMiddleware
+from aiogram.types import TelegramObject
+from sqlalchemy.ext.asyncio import async_sessionmaker
+
+
+class DataBaseSession(BaseMiddleware):
+    def __init__(self, session_pool: async_sessionmaker):
+        self.session_pool = session_pool
+
+
+    async def __call__(self,
+                       handler: Callable[[TelegramObject, dict[str, Any]], Awaitable[Any]],
+                       event: TelegramObject,
+                       data: dict[str, Any]
+                       ) -> Any:
+        async with self.session_pool() as session:
+            data["session"] = session
+            return await handler(event, data)
+    # TODO: разобраться в подходе с разделением чтения и записи, поменять все на конструкцию async with
